updated tests

+++ src/main/java/fr/univavignon/pokedex/api/PokemonFactory.java
@@ -8,9 +8,9 @@ public class PokemonFactory implements IPokemonFactory {
         int defense = (int) (Math.random() * 16);
         int stamina = (int) (Math.random() * 16);
 
-        int totalIV = attack + defense + stamina;
-        int maxIV = 45;
-        int iv = (totalIV / maxIV) * 100;
+        double totalIV = attack + defense + stamina;
+        double maxIV = 45;
+        double iv = (totalIV / maxIV) * 100;
 
         return new Pokemon(index, "NomDuPokemon", attack, defense, stamina, cp, hp, dust, candy, iv);
     }
